# backend/openapi.yaml
openapi: 3.0.0
info:
  title: 'Chiorino ERP API'
  version: '1'
paths:
...
  '/warehouses/{id}/locations':
    post:
      tags:
        - Warehouse
        - masterdata_warehouse_location.php
      summary: 'Create new Warehouse Location'
      operationId: 5f1060e929e8d19c93157e9d68864143
      parameters:
        -
          $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - desc
                - IDwarehouse
              properties:
                desc:
                  type: string
                note:
                  type: string
                IDwarehouse:
                  type: integer
                IDwh_loc_Type:
                  type: integer
              type: object
            example:
              desc: 'Nome destinazione'
              note: note
              IDwarehouse: 1
              IDwh_loc_Type: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseLocationResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
  '/warehouses/{id}/locations/{location_id}':
    get:
      tags:
        - Warehouse
        - masterdata_warehouse_location.php
      summary: 'Warehouse Location show'
      operationId: 0f3d6a236c7b9434c7f1c96056885b6a
      parameters:
        -
          $ref: '#/components/parameters/id'
        -
          $ref: '#/components/parameters/location_id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseLocationResource'
        '404':
          description: 'Not Found'
        '403':
          description: Unauthorized
    put:
      tags:
        - Warehouse
        - masterdata_warehouse_location.php
      summary: 'Update Warehouse Location'
      operationId: a46c895f8298370c62a88fb665652105
      parameters:
        -
          $ref: '#/components/parameters/id'
        -
          $ref: '#/components/parameters/location_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                desc:
                  type: string
              type: object
            example:
              desc: 'Nome destinazione'
              note: note
              IDwh_loc_Type: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseLocationResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
    delete:
      tags:
        - Warehouse
        - masterdata_warehouse_location.php
      summary: 'Delete BP destination'
      operationId: b3efd23d788da13aa24197686cc3fe9f
      parameters:
        -
          $ref: '#/components/parameters/id'
        -
          $ref: '#/components/parameters/location_id'
      responses:
        '200':
          description: OK
        '403':
          description: Unauthorized
        '404':
          description: 'Not Found'
        '500':
          description: 'Internal Error'
...